renamed files containing nft so that they can be used with whichever final contract is chosen
