+++ client/src/contracts/deployments.ts
@@ -3,5 +3,5 @@ export const Goerli = {
 }
 
 export const Rinkeby = {
-  grantNft: '0x1387FC00bf4a351B65d34cCD2acC4e8CB50ef0eF'
+  grantNft: '0xC7B783ea546db45f4bFC1899446F3675A91Dad87'
 }
